Move TLS helpers to a separate lib
