+++ agent/main.go
@@ -14,6 +14,7 @@ import (
 	"time"
 
 	"github.com/jveski/recompose/common"
+	"github.com/jveski/recompose/internal/rpc"
 )
 
 func main() {
@@ -35,7 +36,7 @@ func main() {
 		log.Fatalf("fatal error while creating directory: %s", err)
 	}
 
-	cert, _, err := common.GenCertificate(".")
+	cert, _, err := rpc.GenCertificate(".")
 	if err != nil {
 		log.Fatalf("fatal error while generating certificate: %s", err)
 	}
