+++ common/common_test.go
@@ -2,13 +2,10 @@ package common
 
 import (
 	"context"
-	"os"
-	"path/filepath"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestRunLoop(t *testing.T) {
@@ -78,29 +75,4 @@ func TestStateContainer(t *testing.T) {
 	assert.Equal(t, 0, s.Get())
 	s.Swap(123)
 	assert.Equal(t, 123, s.Get())
-}
-
-func TestGenCertificate(t *testing.T) {
-	dir := t.TempDir()
-
-	// Initial generation
-	cert, fingerprint, err := GenCertificate(dir)
-	require.NoError(t, err)
-	assert.Len(t, fingerprint, 64)
-	assert.Equal(t, fingerprint, GetCertFingerprint(cert.Leaf.Raw))
-	initialFingerprint := fingerprint
-
-	// Load
-	cert, fingerprint, err = GenCertificate(dir)
-	require.NoError(t, err)
-	assert.Equal(t, initialFingerprint, fingerprint)
-	assert.Equal(t, initialFingerprint, GetCertFingerprint(cert.Leaf.Raw))
-
-	// The fingerprint file is regenerated if removed
-	fingerprintPath := filepath.Join(dir, "tls", "cert-fingerprint.txt")
-	require.NoError(t, os.Remove(fingerprintPath))
-	cert, fingerprint, err = GenCertificate(dir)
-	require.NoError(t, err)
-	assert.Equal(t, initialFingerprint, fingerprint)
-	assert.Equal(t, initialFingerprint, GetCertFingerprint(cert.Leaf.Raw))
 }
